fix: fix percent issue

+++ programs/mintedgem/src/events/set_percent.rs
@@ -2,5 +2,5 @@ use anchor_lang::prelude::*;
 
 #[event]
 pub struct SetPercent {
-    pub percent: u64,
+    pub percent: u16,
 }
